Generalisation of utils and queries
