+++ src/graphql/queries.js
@@ -1,0 +1,24 @@
+const login = `mutation Login($credentials: Credentials!) {
+  login(credentials: $credentials) {
+    token
+    message
+    user {
+      id
+      user_name
+      email
+    }
+  }
+}`;
+
+const register = `mutation Register($user: UserInput!) {
+  register(user: $user) {
+    message
+    user {
+      id
+      user_name
+      email
+    }
+  }
+}`;
+
+export { login, register };
