+++ src/js/utils.js
@@ -1,0 +1,13 @@
+const appendAlert = (alertElement, message, type) => {
+  const wrapper = document.createElement("div");
+  wrapper.innerHTML = [
+    `<div class="alert alert-${type} alert-dismissible" role="alert">`,
+    `   <div>${message}</div>`,
+    '   <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>',
+    "</div>",
+  ].join("");
+
+  alertElement.append(wrapper);
+};
+
+export { appendAlert };
