+++ src/login.html
@@ -17,6 +17,7 @@
 
   <title>Login | WikipStudy</title>
   <link rel="icon" href="img/logo.ico" type="image/x-icon" />
+  <script type="module" src="./js/login.js"></script>
 </head>
 
 <body>
@@ -89,7 +90,6 @@
     integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
     crossorigin="anonymous"></script>
 
-  <script src="js/login.js"></script>
 </body>
 
 </html>
