+++ src/register.html
@@ -17,6 +17,7 @@
 
   <title>Register | WikipStudy</title>
   <link rel="icon" href="img/logo.ico" type="image/x-icon" />
+  <script type="module" src="./js/register.js"></script>
 </head>
 
 <body>
@@ -97,7 +98,6 @@
     integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
     crossorigin="anonymous"></script>
 
-  <script src="js/register.js"></script>
 </body>
 
 </html>
